Add notifier script dim-screen2.sh
This script is an example on how to use SIGHUP and SIGTERM signals
sent to the notifier command.

It dims the screen like dim-screen.sh, but 1) instantly restores screen
brightness to the original level when interrupted during fade and 2)
when locking, waits a second before restoring brightness, so that if the
locking command turns of the screen completely, the screen does not
flash to normal brightness just before turning off.